Pedorrada

+++ math/criba.cpp
@@ -23,7 +23,7 @@ void crearcriba(){
 	for(int p=25;p<=MAXP;p+=10) criba[p]=5;
 	for(int p=9;p<=MAXP;p+=6) criba[p]=3; 
 	for(int p=4;p<=MAXP;p+=2) criba[p]=2;
-	for(int p=7,cur=0;(ll)p*p<=MAXP;p+=w[cur++&7]) if (!criba[p]) 
+	for(int p=7,cur=0;p*p<=MAXP;p+=w[cur++&7]) if (!criba[p]) 
 		for(int j=p*p;j<=MAXP;j+=(p<<1)) if(!criba[j]) criba[j]=p;
 }
